GH-37: Changed mapservice to conform to the new MapRepository.

+++ src/EvoSC.Common/Database/DbAccess/Maps/MapRepository.cs
@@ -17,7 +17,7 @@ public class MapRepository : IMapRepository
         _db = db;
     }
 
-    public async Task<DbMap?> GetMapById(int id)
+    public async Task<DbMap?> GetMapById(long id)
     {
         var query = "select * from `Maps` where `Id`=@MapId limit 1";
         return await _db.QueryFirstOrDefaultAsync<DbMap>(query, new
@@ -56,7 +56,7 @@ public class MapRepository : IMapRepository
         return dbMap;
     }
 
-    public async Task<DbMap> UpdateMap(int mapId, Map map)
+    public async Task<DbMap> UpdateMap(long mapId, Map map)
     {
         var updatedMap = new DbMap
         {
